requirements.txt
slack_bolt==1.20.0
Flask==2.2.5
python-dotenv==1.0.0
SQLAlchemy==1.4.46
APScheduler==3.10.4
gunicorn==20.1.0
requests==2.31.0